Added quotation marks

+++ hypervm-install/hypervm-linux/hypervm-install-slave.sh
@@ -29,41 +29,43 @@ cat LICENSE | more
 echo "--------------------------------------------"
 
 start() {
-
 	export PATH=/usr/sbin:/sbin:$PATH
 
-if ! [ -f /usr/bin/yum ] ; then
-      	echo You at least need yum installed for this to work...
-	echo Please contact us or visit the forum at http://forum.lxcenter.org
-	echo "                                "
-	exit
-fi
+	if ! [ -f /usr/bin/yum ] ; then
+     	echo "You at least need yum installed for this to work..."
+		echo "Please contact us or visit the forum at http://forum.lxcenter.org"
+		echo "                                "
+		exit
+	fi
 #
-if [ -f /usr/bin/yum ] ; then
-	echo Installing some packages with yum
-	yum -y install php wget zip unzip 
-else 
-	echo Installing some packages with up2date
-	up2date --nox --nosig php wget zip unzip
-fi
-#
+	echo "Installing some packages with yum"
+	yum -y install php wget zip unzip
+
 	echo Checking if php is installed
-if ! [ -f /usr/bin/php ] ; then
-	echo Installing php failed. Please fix yum/up2date.
-	exit
-fi
-#
-if 	[ -f ./hypervm-install.zip ] ; then
-	echo Remove old installation package
-	rm -f hypervm-install.zip
-fi
-	echo Downloading installation package from LxCenter
-	wget http://download.lxcenter.org/download/hypervm-install.zip
-#
-	echo Unpacking installation package	
-	unzip -oq hypervm-install.zip
+	if ! [ -f /usr/bin/php ] ; then
+		echo "Installing php failed. Please fix yum."
+		exit
+	fi
+
+	if [ ! -d '../../.git' ]; then
+
+		if [ -f ./hypervm-install.zip ] ; then
+    		echo "Remove old installation package"
+    		rm -f hypervm-install.zip
+    	fi
+		pwd
+		echo "Downloading installation package from LxCenter"
+		wget http://download.lxcenter.org/download/hypervm-install.zip
+		echo "Unpacking installation package"
+		unzip -oq hypervm-install.zip
+	else
+		echo "Development GIT version found. Skipping download sources."
+		echo "Unpacking installation package from current development version."
+		unzip -oq ../hypervm-install.zip
+	fi
+
 	cd hypervm-install/hypervm-linux
-	echo Starting main installation script
+	echo "Starting main installation script"
 	php lxins.php --install-type=slave $* | tee hypervm_install.log
 }
 #
@@ -71,18 +73,21 @@ fi
 #
 case "$1" in
   --virtualization-type=xen)
+	echo 'Installing HyperVM with Xen virtualization'
     start $*
     ;;
   --virtualization-type=openvz)
+	echo 'Installing HyperVM with OpenVZ virtualization'
     start $*
     ;;
   --virtualization-type=NONE)
+	echo 'Installing HyperVM with default(Xen) virtualization'
     start $*
     ;;
   *)
-   	echo $"This is the HyperVM Install script"
-    	echo $"The usage is:"
-    	echo $"sh $0 --virtualization-type=xen/openvz/NONE [--skipostemplate=true]"
+    echo $"This is the HyperVM Install script"
+    echo $"The usage is:"
+    echo $"sh $0 --virtualization-type=xen/openvz/NONE [--skipostemplate=true]"
 	exit 1
 esac
 exit $?
